+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [v8.23.0] - 2026-08-14
+
 ### Changed
 
 - Minor `QueryManager` behavior changes:
@@ -2506,7 +2508,8 @@ _(This list may look long, but the breaking changes should only affect a small m
 
 <!-- #region Release comparison links -->
 
-[unreleased]: https://github.com/react-querybuilder/react-querybuilder/compare/v8.22.5...HEAD
+[unreleased]: https://github.com/react-querybuilder/react-querybuilder/compare/v8.23.0...HEAD
+[v8.23.0]: https://github.com/react-querybuilder/react-querybuilder/compare/v8.22.4...v8.23.0
 [v8.22.5]: https://github.com/react-querybuilder/react-querybuilder/compare/v8.22.4...v8.22.5
 [v8.22.4]: https://github.com/react-querybuilder/react-querybuilder/compare/v8.22.3...v8.22.4
 [v8.22.3]: https://github.com/react-querybuilder/react-querybuilder/compare/v8.22.2...v8.22.3